# blueprints/job-saas/worker/job/routes.ts
import { Hono } from "hono";
import type { Env } from "../env";
import { requireAuth } from "../middleware";
import { careerFacts, recordAiUsage, upsertJob } from "./store";
import { runDiscovery } from "./discovery";
import { sourceAdapter } from "./sources";
import { safeJson, now } from "./util";
import { ingestLinkedInAlertText, ingestLinkedInUrl } from "./linkedin";
import { autopilotDecision } from "./application-guard";
import type { NormalizedJob } from "./types";
import { scoreAndPersist, prepareAndPersist } from "./service";
import { runUserCycle } from "./cycle";
import { draftQuestionAnswer } from "./ai/questions";
import { extractCareerFile } from "./ai/career";

export const jobRoutes=new Hono<{Bindings:Env;Variables:{session:any}}>();
jobRoutes.use("/*",requireAuth as any);
const uid=(c:any)=>String(c.get("session").user.id);

jobRoutes.get("/career",async c=>{const facts=await c.env.DB.prepare("SELECT id,type,employer,role,text,verification,source,updated_at FROM career_facts WHERE user_id=? ORDER BY updated_at DESC").bind(uid(c)).all();return c.json({facts:facts.results})});

function base64Bytes(bytes:Uint8Array):string{let binary="";for(let i=0;i<bytes.length;i+=0x8000)binary+=String.fromCharCode(...bytes.subarray(i,Math.min(i+0x8000,bytes.length)));return btoa(binary)}
jobRoutes.post("/career/extract-file/:fileId",async c=>{const userId=uid(c);const row=await c.env.DB.prepare("SELECT id,r2_key,name,size,content_type FROM files WHERE id=? AND user_id=?").bind(c.req.param("fileId"),userId).first<any>();if(!row)return c.json({error:"File not found"},404);if(Number(row.size)>5*1024*1024)return c.json({error:"CV extraction limit is 5 MB"},413);const allowed=["application/pdf","application/vnd.openxmlformats-officedocument.wordprocessingml.document","text/plain"];if(!allowed.includes(row.content_type))return c.json({error:"Use PDF, DOCX or TXT for CV extraction"},415);const obj=await c.env.FILES.get(row.r2_key);if(!obj)return c.json({error:"Object missing"},404);const bytes=new Uint8Array(await obj.arrayBuffer());const r=await extractCareerFile(c.env,{filename:row.name,contentType:row.content_type,base64:base64Bytes(bytes)});const ts=now();await c.env.DB.prepare("INSERT INTO career_profiles(user_id,headline,summary,default_resume_file_id,completeness,updated_at) VALUES(?,?,?,?,?,?) ON CONFLICT(user_id) DO UPDATE SET headline=excluded.headline,summary=excluded.summary,default_resume_file_id=excluded.default_resume_file_id,updated_at=excluded.updated_at").bind(userId,r.data.headline||"",r.data.summary||"",row.id,0,ts).run();let inserted=0;for(const f of (r.data.facts||[]).slice(0,200)){if(!f.text)continue;await c.env.DB.prepare("INSERT INTO career_facts(id,user_id,type,employer,role,text,verification,source,source_file_id,created_at,updated_at) VALUES(?,?,?,?,?,?,'inferred','cv_extract',?,?,?)").bind(crypto.randomUUID(),userId,f.type||"other",f.employer||null,f.role||null,String(f.text),row.id,ts,ts).run();inserted++}await recordAiUsage(c.env,{userId,operation:"career_extract",model:r.model,inputTokens:r.inputTokens,outputTokens:r.outputTokens,cost:r.estimatedCostUsd});return c.json({headline:r.data.headline,summary:r.data.summary,factsInserted:inserted,verification:"inferred"})});

jobRoutes.post("/career/facts",async c=>{const b=await c.req.json<any>();if(!b.text||!b.type)return c.json({error:"type and text required"},400);const id=crypto.randomUUID(),ts=now();const verification=["verified","user_provided","inferred","unverified"].includes(b.verification)?b.verification:"user_provided";await c.env.DB.prepare("INSERT INTO career_facts(id,user_id,type,employer,role,text,verification,source,created_at,updated_at) VALUES(?,?,?,?,?,?,?,?,?,?)").bind(id,uid(c),b.type,b.employer||null,b.role||null,String(b.text),verification,b.source||"user",ts,ts).run();return c.json({id,verification},201)});

jobRoutes.put("/career/facts/:id",async c=>{const b=await c.req.json<any>();const verification=["verified","user_provided","inferred","unverified"].includes(b.verification)?b.verification:null;if(!verification)return c.json({error:"Invalid verification"},400);const r=await c.env.DB.prepare("UPDATE career_facts SET verification=?,updated_at=? WHERE id=? AND user_id=?").bind(verification,now(),c.req.param("id"),uid(c)).run();if(!r.meta.changes)return c.json({error:"Not found"},404);return c.json({saved:true,verification})});
jobRoutes.post("/career/confirm",async c=>{const b=await c.req.json<any>();const ids=Array.isArray(b.ids)?b.ids.map(String).slice(0,200):[];if(!ids.length)return c.json({error:"Select at least one extracted fact to confirm"},400);let confirmed=0;for(const id of ids){const r=await c.env.DB.prepare("UPDATE career_facts SET verification='verified',updated_at=? WHERE id=? AND user_id=? AND verification='inferred'").bind(now(),id,uid(c)).run();confirmed+=Number(r.meta.changes||0)}return c.json({confirmed})});

jobRoutes.get("/searches",async c=>{const r=await c.env.DB.prepare("SELECT * FROM job_searches WHERE user_id=? ORDER BY created_at DESC").bind(uid(c)).all();return c.json(r.results)});
jobRoutes.post("/searches",async c=>{const b=await c.req.json<any>();const id=crypto.randomUUID(),ts=now();await c.env.DB.prepare(`INSERT INTO job_searches(id,user_id,name,roles_json,locations_json,industries_json,excluded_companies_json,excluded_keywords_json,remote_preference,minimum_salary,salary_currency,minimum_match,max_applications_per_day,auto_prepare,auto_submit,created_at,updated_at) VALUES(?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)`).bind(id,uid(c),b.name||"My Job Agent",JSON.stringify(b.roles||[]),JSON.stringify(b.locations||[]),JSON.stringify(b.industries||[]),JSON.stringify(b.excludedCompanies||[]),JSON.stringify(b.excludedKeywords||[]),b.remotePreference||null,b.minimumSalary??null,b.salaryCurrency||null,Number(b.minimumMatch||85),Math.max(1,Math.min(50,Number(b.maxApplicationsPerDay||5))),b.autoPrepare===false?0:1,b.autoSubmit?1:0,ts,ts).run();return c.json({id},201)});

jobRoutes.get("/sources",async c=>{const r=await c.env.DB.prepare("SELECT * FROM job_source_configs WHERE user_id=? OR user_id IS NULL ORDER BY provider,label").bind(uid(c)).all();return c.json(r.results)});
jobRoutes.post("/sources",async c=>{const b=await c.req.json<any>();if(!sourceAdapter(b.provider))return c.json({error:"Unsupported built-in provider"},400);const id=crypto.randomUUID();await c.env.DB.prepare("INSERT OR IGNORE INTO job_source_configs(id,user_id,provider,identifier,label,enabled,metadata_json,created_at) VALUES(?,?,?,?,?,1,'{}',?)").bind(id,uid(c),b.provider,String(b.identifier||""),b.label||b.identifier||b.provider,now()).run();return c.json({id},201)});
jobRoutes.post("/discover",async c=>c.json(await runUserCycle(c.env,uid(c))));

jobRoutes.get("/jobs",async c=>{const min=Number(c.req.query("minScore")||0);const r=await c.env.DB.prepare(`SELECT j.*,jm.id match_id,jm.score,jm.strengths_json,jm.gaps_json,jm.explanation FROM jobs j LEFT JOIN job_matches jm ON j.id=jm.job_id AND jm.user_id=? WHERE jm.score IS NULL OR jm.score>=? ORDER BY CASE WHEN jm.score IS NULL THEN 1 ELSE 0 END,jm.score DESC,j.discovered_at DESC LIMIT 200`).bind(uid(c),min).all<any>();return c.json(r.results.map(x=>({...x,strengths:safeJson(x.strengths_json,[]),gaps:safeJson(x.gaps_json,[])})))});
jobRoutes.get("/jobs/:id",async c=>{const j=await c.env.DB.prepare("SELECT * FROM jobs WHERE id=?").bind(c.req.param("id")).first();if(!j)return c.json({error:"Not found"},404);const m=await c.env.DB.prepare("SELECT * FROM job_matches WHERE user_id=? AND job_id=?").bind(uid(c),c.req.param("id")).first<any>();return c.json({job:j,match:m?{...m,strengths:safeJson(m.strengths_json,[]),gaps:safeJson(m.gaps_json,[])}:null})});
jobRoutes.post("/jobs/:id/score",async c=>{const body=await c.req.json<any>().catch(()=>({}));try{return c.json(await scoreAndPersist(c.env,uid(c),c.req.param("id"),body.searchId||null))}catch(e){return c.json({error:String(e)},400)}});
jobRoutes.post("/jobs/:id/prepare",async c=>{try{return c.json(await prepareAndPersist(c.env,uid(c),c.req.param("id")))}catch(e){return c.json({error:String(e)},400)}});

jobRoutes.get("/applications",async c=>{const r=await c.env.DB.prepare(`SELECT a.*,j.company,j.title,j.location,j.apply_url,jm.score,(SELECT p.id FROM application_submission_proofs p WHERE p.application_id=a.id ORDER BY p.captured_at DESC LIMIT 1) proof_id FROM applications a JOIN jobs j ON j.id=a.job_id LEFT JOIN job_matches jm ON jm.id=a.match_id WHERE a.user_id=? ORDER BY a.updated_at DESC`).bind(uid(c)).all();return c.json(r.results)});
jobRoutes.get("/applications/:id/proof",async c=>{const p=await c.env.DB.prepare(`SELECT p.r2_key,p.content_type FROM application_submission_proofs p JOIN applications a ON a.id=p.application_id WHERE p.application_id=? AND a.user_id=? ORDER BY p.captured_at DESC LIMIT 1`).bind(c.req.param("id"),uid(c)).first<any>();if(!p)return c.json({error:"Submission proof not available"},404);const obj=await c.env.FILES.get(p.r2_key);if(!obj)return c.json({error:"Proof object missing"},404);return new Response(obj.body,{headers:{"content-type":p.content_type,"cache-control":"private, no-store","content-disposition":"inline"}})});
jobRoutes.post("/applications/:id/outcome",async c=>{const userId=uid(c),id=c.req.param("id"),b=await c.req.json<any>();const outcome=String(b.outcome||"");if(!["viewed","interview","offer","rejected","withdrawn"].includes(outcome))return c.json({error:"Invalid outcome"},400);const app=await c.env.DB.prepare("SELECT status FROM applications WHERE id=? AND user_id=?").bind(id,userId).first<{status:string}>();if(!app)return c.json({error:"Not found"},404);const ts=Number(b.occurredAt)||now();await c.env.DB.prepare("INSERT INTO job_outcomes(id,application_id,user_id,outcome,source,detail_json,occurred_at,created_at) VALUES(?,?,?,?,?,?,?,?)").bind(crypto.randomUUID(),id,userId,outcome,"user",JSON.stringify(b.detail||{}),ts,now()).run();if(["interview","offer","rejected","withdrawn"].includes(outcome)){await c.env.DB.prepare("UPDATE applications SET status=?,updated_at=? WHERE id=? AND user_id=?").bind(outcome,now(),id,userId).run();await c.env.DB.prepare("INSERT INTO application_events(id,application_id,from_status,to_status,actor,detail_json,created_at) VALUES(?,?,?,?,?,?,?)").bind(crypto.randomUUID(),id,app.status,outcome,"user",JSON.stringify(b.detail||{}),now()).run()}return c.json({saved:true,outcome})});

jobRoutes.get("/applications/:id/questions",async c=>{const r=await c.env.DB.prepare("SELECT id,question,answer,source,requires_user_input FROM application_answers WHERE application_id=? AND application_id IN (SELECT id FROM applications WHERE user_id=?) ORDER BY created_at").bind(c.req.param("id"),uid(c)).all();return c.json(r.results)});
jobRoutes.post("/applications/:id/questions/:answerId/draft",async c=>{const row=await c.env.DB.prepare("SELECT aa.id,aa.question,a.job_id FROM application_answers aa JOIN applications a ON a.id=aa.application_id WHERE aa.id=? AND a.id=? AND a.user_id=?").bind(c.req.param("answerId"),c.req.param("id"),uid(c)).first<any>();if(!row)return c.json({error:"Not found"},404);const facts=await careerFacts(c.env,uid(c));const r:any=await draftQuestionAnswer(c.env,facts,row.question);if(r.data.answer){await c.env.DB.prepare("UPDATE application_answers SET answer=?,source='ai_draft',requires_user_input=1,updated_at=? WHERE id=?").bind(r.data.answer,now(),row.id).run()}if(r.model!=="deterministic")await recordAiUsage(c.env,{userId:uid(c),jobId:row.job_id,applicationId:c.req.param("id"),operation:"application_answer",model:r.model,inputTokens:r.inputTokens,outputTokens:r.outputTokens,cost:r.estimatedCostUsd});return c.json(r.data)});
jobRoutes.put("/applications/:id/questions/:answerId",async c=>{const b=await c.req.json<any>();const r=await c.env.DB.prepare("UPDATE application_answers SET answer=?,source='user',requires_user_input=0,updated_at=? WHERE id=? AND application_id=? AND application_id IN (SELECT id FROM applications WHERE user_id=?)").bind(String(b.answer||""),now(),c.req.param("answerId"),c.req.param("id"),uid(c)).run();if(!r.meta.changes)return c.json({error:"Not found"},404);const unresolved=await c.env.DB.prepare("SELECT COUNT(*) n FROM application_answers WHERE application_id=? AND requires_user_input=1").bind(c.req.param("id")).first<{n:number}>();await c.env.DB.prepare("UPDATE applications SET unresolved_questions=?,status=CASE WHEN ?=0 AND status='needs_input' THEN 'package_ready' ELSE status END,updated_at=? WHERE id=? AND user_id=?").bind(Number(unresolved?.n||0),Number(unresolved?.n||0),now(),c.req.param("id"),uid(c)).run();return c.json({saved:true,unresolved:Number(unresolved?.n||0)})});

jobRoutes.post("/applications/:id/queue",async c=>{const id=c.req.param("id"),userId=uid(c);const a=await c.env.DB.prepare("SELECT status FROM applications WHERE id=? AND user_id=?").bind(id,userId).first<{status:string}>();if(!a)return c.json({error:"Not found"},404);const body=await c.req.json<any>().catch(()=>({}));if(body.autopilot){const d=await autopilotDecision(c.env,userId,id);if(!d.allowed)return c.json({error:"Autopilot guardrail blocked submission",reasons:d.reasons},409)}if(!["package_ready","ready_for_review","failed"].includes(a.status))return c.json({error:`Cannot queue from ${a.status}`},409);await c.env.DB.prepare("UPDATE applications SET status='queued',submit_mode=?,queued_at=?,updated_at=? WHERE id=? AND user_id=?").bind(body.autopilot?"autopilot":"review",now(),now(),id,userId).run();await c.env.APPLICATION_QUEUE.send({applicationId:id,userId,jobId:String((await c.env.DB.prepare("SELECT job_id FROM applications WHERE id=?").bind(id).first<any>())?.job_id),attempt:0});return c.json({queued:true})});

jobRoutes.post("/ingest/linkedin-url",async c=>{const b=await c.req.json<any>();try{return c.json(await ingestLinkedInUrl(c.env,uid(c),String(b.url||"")),201)}catch(e){return c.json({error:String(e)},400)}});
jobRoutes.post("/ingest/linkedin-alert",async c=>{const b=await c.req.json<any>();return c.json(await ingestLinkedInAlertText(c.env,uid(c),String(b.text||"")),201)});
jobRoutes.post("/ingest/job",async c=>{const b=await c.req.json<NormalizedJob>();if(!b.sourceKey||!b.sourceUrl||!b.company||!b.title||!b.description)return c.json({error:"Missing normalized job fields"},400);return c.json({id:await upsertJob(c.env,{...b,sourceProvider:b.sourceProvider||"manual"})},201)});